% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2b.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate a confidence interval for the sample mean}
\usage{
calculate_CI(x, conf = 0.95)
}
\arguments{
\item{x}{A numeric vector}

\item{conf}{Confidence level (default = 0.95)}
}
\value{
A named numeric vector of length 2:
\itemize{
\item lower_bound
\item upper_bound
}
}
\description{
Calculate a confidence interval for the sample mean
}
\examples{
calculate_CI(c(1, 2, 3, 4, 5), conf = 0.95)
calculate_CI(rnorm(10), conf = 0.9)
}
